afalg: fix AF_ALG AES driver can't receive large data
When we input a large to do aes-256-cbc via AF_ALG interface, we will
encounter the below error.

ALG_ERR: afalg_do_cipher: sent 0x2c000 bytes != len 0x2d000
aes-256-cbc error!

There are two issues in the error. The first one is the send buffer
size is not enough and the other one is the memory data didn't fit
the memory page. Finally, it causes the Kernel drops some bytes that
are not fitted with the AES block size. This code change is used to
fix the send buffer size issue via the setsockopt API.

The maximum value of the send buffer is twice wmem
(/proc/sys/net/core/wmem_max). If an application wants to send a
bigger data to Kernel driver, wmem should be enlarged as well.

Signed-off-by: Kane-Chen-AS <kane_chen@aspeedtech.com>
Change-Id: I7ead254a20f0091f78173ba104d162f5065cc2e2